requirements.txt
appdirs==1.4.3
bottlenose==1.1.2
dj-database-url==0.4.2
Django==1.10.6
lxml==3.7.3
packaging==16.8
pyparsing==2.2.0
python-amazon-simple-product-api==2.2.11
python-dateutil==2.6.0
six==1.10.0
whitenoise==3.3.0
